Feat: review의 rating(별점)필드 int로 수정

+++ core/src/main/kotlin/elice/team5th/domain/review/dto/ReviewDTO.kt
@@ -5,7 +5,7 @@ data class ReviewDTO(
     val userId: Long,
     val contentId: Long,
     val detail: String,
-    val rating: Double,
+    val rating: Int,
     val likes: Int,
     val reports: Int
 )
